# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Facebook Ads Marketing API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/product/list": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Product List",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "description": "status",
                                    "type": "string"
                                },
                                "data": {
                                    "description": "The response data",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Product"
                                    }
                                },
                                "message": {
                                    "description": "Message",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create Product",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Product Name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "description": "status",
                                    "type": "string"
                                },
                                "data": {
                                    "description": "The response data",
                                    "type": "object",
                                    "$ref": "#/definitions/Product"
                                },
                                "message": {
                                    "description": "Message",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/product/{productID}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Product by id",
                "parameters": [
                    {
                        "name": "productID",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "description": "status",
                                    "type": "string"
                                },
                                "data": {
                                    "description": "The response data",
                                    "type": "object",
                                    "$ref": "#/definitions/Product"
                                },
                                "message": {
                                    "description": "Message",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update Product by id",
                "parameters": [
                    {
                        "name": "productID",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Product Name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "description": "status",
                                    "type": "string"
                                },
                                "data": {
                                    "description": "The response data",
                                    "type": "object",
                                    "$ref": "#/definitions/Product"
                                },
                                "message": {
                                    "description": "Message",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product by id",
                "parameters": [
                    {
                        "name": "productID",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "description": "status",
                                    "type": "string"
                                },
                                "data": {
                                    "description": "The response data",
                                    "type": "object",
                                    "$ref": "#/definitions/Product"
                                },
                                "message": {
                                    "description": "Message",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/product/{id}/ad_account/list": {
            "get": {
                "tags": [
                    "Ad Account"
                ],
                "summary": "Get Ad Accounts by Product ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "description": "status",
                                    "type": "string"
                                },
                                "data": {
                                    "description": "The response data",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/AdAccount"
                                    }
                                },
                                "message": {
                                    "description": "Message",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "AdAccount": {
            "description": "Response for Ad Account",
            "properties": {
                "id": {
                    "description": "UUID",
                    "type": "integer"
                },
                "user_id": {
                    "description": "User UUID",
                    "type": "integer"
                },
                "fb_user_id": {
                    "description": "FB User ID",
                    "type": "string"
                },
                "product_id": {
                    "description": "Product UUID",
                    "type": "integer | null"
                },
                "ad_account_id": {
                    "description": "FB Ad Account Name",
                    "type": "string"
                },
                "ad_account_name": {
                    "description": "FB Ad Account ID",
                    "type": "string"
                },
                "created_at": {
                    "description": "Created Date",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Updated Date",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "Product": {
            "description": "Response for Product",
            "properties": {
                "id": {
                    "description": "Product UUID",
                    "type": "integer"
                },
                "user_id": {
                    "description": "User UUID",
                    "type": "integer"
                },
                "name": {
                    "description": "FB User ID",
                    "type": "string"
                },
                "created_at": {
                    "description": "Created Date",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Updated Date",
                    "type": "string"
                }
            },
            "type": "object"
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}